+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 ### Fixed
 - Camera now maintains own z position instead of adopting the players
 - Fixed off-by-one error in collision tile placement
+- Camera now updates position correctly upon changing window size
 
 ## [0.2.0] - 2022-07-28
 ### Added